% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance_kproto}
\alias{importance_kproto}
\title{Variable Importance}
\usage{
importance_kproto(object, rand = TRUE)
}
\arguments{
\item{object}{Object of class \code{kproto}.}

\item{rand}{Logical indicating whether additional variable importance based on rand indices should be computed.}
}
\value{
List of three elements:

\item{vi_decomp}{Data frame where the rows are: the contribution of the variables to \code{object$tot.withinss}, 
                         the the contribution of the variables in a null model, the differences and releatiove differences}

\item{vi_rand}{Vector of variable importances based on the adjusted Rand index after reclustering.}
}
\description{
Computes two types of variable importance for an object of class \code{kproto}.
}
\details{
The contribution of the variables to \code{object$tot.withinss} is computed 
as well as those for a null model where all variables are in the same cluster. 
Variable importance is computed as variable-wise improvement of \code{object$tot.withinss}.

As proposed by Henning and Murphy (2023) the importance of a variable can also be assessed by 
comparing how much the resulting clusters will change if a variable is ignored: 
For \code{rand == TRUE} in addition variable importance is computed by reclustering the data 
where the variable of interest is replaced by a constant value.
Variable importance (VI) is computed by comparing the original partition 
with the one obtained after reclustering the data using adjusted Rand indices as \eqn{VI = 1-Rand_{adj}}. 

As reclustering for each variable may be time consuming, it may be descelected.

Note that for \code{na.rm = "imp.onestep"} and \code{"imp.internal"} the imputed data are used. 
For \code{na.rm = "yes"} observations with missing values are removed from the data and 
for \code{na.rm = "no"} \code{NA}s are ignored for distance computation just as in \code{kproto()}.
}
\examples{
# generate toy data with factors and numerics

n   <- 100
prb <- 0.9
muk <- 1.5 
clusid <- rep(1:4, each = n)

x1 <- sample(c("A","B"), 2*n, replace = TRUE, prob = c(prb, 1-prb))
x1 <- c(x1, sample(c("A","B"), 2*n, replace = TRUE, prob = c(1-prb, prb)))
x1 <- as.factor(x1)

x2 <- sample(c("A","B"), 2*n, replace = TRUE, prob = c(prb, 1-prb))
x2 <- c(x2, sample(c("A","B"), 2*n, replace = TRUE, prob = c(1-prb, prb)))
x2 <- as.factor(x2)

x3 <- c(rnorm(n, mean = -muk), rnorm(n, mean = muk), rnorm(n, mean = -muk), rnorm(n, mean = muk))
x4 <- c(rnorm(n, mean = -muk), rnorm(n, mean = muk), rnorm(n, mean = -muk), rnorm(n, mean = muk))

x <- data.frame(x1,x2,x3,x4)

# apply k-prototypes
kpres <- kproto(x, 4)

# compute different kinds of variable importance
importance_kproto(kpres, rand = TRUE)

# reducing the influence of the categorical variables (X1 and X2) on the clustering:
kpres <- kproto(x, 4, lambda = 0.01)
importance_kproto(kpres, rand = FALSE)

# reducing the importance of the numeric variables (X3 and X4) on the clustering:
kpres <- kproto(x, 4, lambda = 100)
importance_kproto(kpres, rand = FALSE)

# weighting all variables equally:
kpres <- kproto(x, 4, lambda = 1)
importance_kproto(kpres, rand = FALSE)

}
\references{
\itemize{
    \item Szepannek, G. (2018):
    clustMixType: User-Friendly Clustering of Mixed-Type Data in R 
    {\emph{The R Journal 10/2}}, 200-208,  
    \doi{10.32614/RJ-2018-048}.
    \item Hennig, C., and K. Murphy (2023): 
    Quantyfying Variable Importance in Cluster Analysis, 
    {\emph{CLADAG} 2023 Book of Abstracts and Short Papers}, 515-518,  
    Pearson Education Resources.
    \item Aschenbruck, R., Szepannek, G., Wilhelm, A. (2022): 
    Imputation Strategies for Clustering Mixed‑Type Data with Missing Values, 
    {\emph{Journal of Classification}}, 
    \doi{10.1007/s00357-022-09422-y}. 
    }.
}
\author{
\email{gero.szepannek@web.de}
}
